Refactor tests into separate modules
