+++ tests/test_config.py
@@ -1,0 +1,26 @@
+import pytest
+
+import src.main as main
+
+
+def test_load_config_success(tmp_path, monkeypatch):
+    cfg_file = tmp_path / "cfg.yml"
+    cfg_file.write_text("foo: 1")
+    monkeypatch.setattr(main, "CONFIG_PATH", str(cfg_file))
+    assert main.load_config() == {"foo": 1}
+
+
+def test_load_config_missing(tmp_path, monkeypatch):
+    monkeypatch.setattr(main, "CONFIG_PATH", str(tmp_path / "nonexistent.yml"))
+    with pytest.raises(FileNotFoundError):
+        main.load_config()
+
+
+def test_get_api_credentials_success():
+    cfg = {"api_id": "123", "api_hash": "hash", "session": "sess"}
+    assert main.get_api_credentials(cfg) == (123, "hash", "sess")
+
+
+def test_get_api_credentials_missing():
+    with pytest.raises(RuntimeError):
+        main.get_api_credentials({})
